Introduce DRM / GBM renderer
This is a variation of the EGL renderers that uses GBM to create
buffers directly. The GBM renderer creates a double-buffered surface,
then bridges to OpenGL through the EGL GBM platform.

Except for opening up a GBM device from the adapter number, and the
direct buffer management, this still looks a lot like a regular EGL
renderer.

As it is, this is not useful yet; next commits should introduce a
mechanism for these buffers to be exported and shared, potentially
across processes.